re-organisation of code
